change place cards

+++ app/src/main/java/com/mikazil/samsung_project/HourlyForecast.java
@@ -2,11 +2,11 @@ package com.mikazil.samsung_project;
 
 public class HourlyForecast {
     private final String time;
-    private final int temperature;
+    private final double temperature; // Изменено с int на double
     private final int iconRes;
     private final boolean isNow;
 
-    public HourlyForecast(String time, int temperature, int iconRes, boolean isNow) {
+    public HourlyForecast(String time, double temperature, int iconRes, boolean isNow) {
         this.time = time;
         this.temperature = temperature;
         this.iconRes = iconRes;
@@ -14,7 +14,7 @@ public class HourlyForecast {
     }
 
     public String getTime() { return time; }
-    public double getTemperature() { return temperature; }
+    public double getTemperature() { return temperature; } // Возвращает double
     public int getIconRes() { return iconRes; }
     public boolean isNow() { return isNow; }
 }
